Move stuff from sct_maths to sct.math + cleanup imports
